requirements.txt
click==8.1.7
typer==0.12.5
typing_extensions==4.12.2
torch==2.4.1
numpy==2.1.1
kafka-python-ng==2.2.3
